[Feat]: 코드 스멜 제거

+++ src/main/java/shop/yesaladin/shop/order/service/impl/QueryOrderCouponServiceImpl.java
@@ -61,7 +61,7 @@ public class QueryOrderCouponServiceImpl implements QueryOrderCouponService {
         //상품의 판매가
         long saleAmount = Math.round((product.getActualPrice() * request.getQuantity()) *
                 (100 - (product.isSeparatelyDiscount() ? product.getDiscountRate()
-                        : product.getTotalDiscountRate().getDiscountRate())) / 1000) * 10L;
+                        : product.getTotalDiscountRate().getDiscountRate())) / (float) 1000) * 10L;
 
         //상품의 실판매가
         long couponAppliedAmount = getCouponAppliedAmount(
